requirements.txt
google
google-genai
python-dotenv
pandas
openpyxl